03-kafka-serde-starter/kafka-serde-starter/requirements.txt
confluent-kafka[json,avro,protobuf]==1.7.0
PyYAML==5.4.1
